+++ resources/js/app.js
@@ -1,8 +1,6 @@
 import Alpine from 'alpinejs';
 import Focus from '@alpinejs/focus';
-
-import './clipboard';
-import './components/search';
+import docsearch from '@docsearch/js';
 
 import.meta.glob([
   '../images/**',
@@ -13,10 +11,12 @@ window.Alpine = Alpine;
 Alpine.plugin(Focus);
 Alpine.start();
 
-document.addEventListener('DOMContentLoaded', () => {
-    if (document.querySelector('#docsScreen')) {
-        import('./docs.js');
-    }
-
-    import('./components/accessibility');
+docsearch({
+    container: '#docsearch',
+    appId: algolia_app_id,
+    apiKey: algolia_search_key,
+    indexName: 'laravel',
+    searchParameters: {
+        facetFilters: ['version:' + window.version],
+    },
 });
